requirements/base.txt
# testing
pytest
pytest-django

# Factory boy (https://factoryboy.readthedocs.io/en/latest/)
factory_boy

# Rest framework (https://www.django-rest-framework.org/)
djangorestframework

# Django money (https://github.com/django-money/django-money)
django-money


# Django Filter (https://django-filter.readthedocs.io/en/latest/guide/usage.html#the-filter)
django-filter

# Graphql extensions (https://github.com/flavors/django-graphql-extensions)
django-graphql-extensions

# quiz (https://github.com/ariebovenberg/quiz)
quiz

# requests (https://requests.kennethreitz.org/en/master/)
requests